Removing unused stuff.
Tests are cool, and everybody loves some TTD - but
when shit hits the fan, I don't have the time to
fiddle around with them. Maybe I'll find a better,
more agile solution to this problem later. For now
I'll just can it.

Also, scripts are a good idea - but the current
implementation simply sucks.